backend/requirements.txt
DJango>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
